IC-2060: Change view appointment feedback classes to be agnostic to appointment type (actionplan vs supplier assessment)
