+++ server/routes/shared/appointment/feedback/submittedFeedbackPresenter.ts
@@ -1,0 +1,69 @@
+import DeliusServiceUser from '../../../../models/delius/deliusServiceUser'
+import User from '../../../../models/hmppsAuth/user'
+import { ActionPlanAppointment } from '../../../../models/actionPlan'
+import DateUtils from '../../../../utils/dateUtils'
+import { SummaryListItem } from '../../../../utils/summaryList'
+import FeedbackAnswersPresenter from '../../../service-provider/appointment/feedback/feedbackAnswersPresenter'
+import AttendanceFeedbackPresenter from '../../../service-provider/appointment/feedback/attendance/attendanceFeedbackPresenter'
+import BehaviourFeedbackPresenter from '../../../service-provider/appointment/feedback/behaviour/behaviourFeedbackPresenter'
+import ActionPlanPostSessionAttendanceFeedbackPresenter from '../../../service-provider/action-plan/appointment/post-session-feedback/attendance/actionPlanPostSessionAttendanceFeedbackPresenter'
+import Appointment from '../../../../models/appointment'
+import InitialAssessmentAttendanceFeedbackPresenter from '../../../service-provider/referrals/supplier-assessment/post-assessment-feedback/attendance/initialAssessmentAttendanceFeedbackPresenter'
+
+export default class SubmittedFeedbackPresenter extends FeedbackAnswersPresenter {
+  protected readonly attendancePresenter: AttendanceFeedbackPresenter
+
+  protected readonly behaviourPresenter: BehaviourFeedbackPresenter
+
+  constructor(
+    appointmentDetails: ActionPlanAppointment | Appointment,
+    private readonly serviceUser: DeliusServiceUser,
+    private readonly assignedCaseworker: User | null = null
+  ) {
+    super(appointmentDetails)
+    if (this.isActionPlanAppointment(appointmentDetails)) {
+      this.attendancePresenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(
+        appointmentDetails,
+        this.serviceUser
+      )
+    } else {
+      this.attendancePresenter = new InitialAssessmentAttendanceFeedbackPresenter(appointmentDetails, this.serviceUser)
+    }
+    this.behaviourPresenter = new BehaviourFeedbackPresenter(appointmentDetails, this.serviceUser)
+  }
+
+  private isActionPlanAppointment(
+    appointmentDetails: Appointment | ActionPlanAppointment
+  ): appointmentDetails is ActionPlanAppointment {
+    return (<ActionPlanAppointment>appointmentDetails).sessionNumber !== undefined
+  }
+
+  readonly text = {
+    title: `View feedback`,
+  }
+
+  get sessionDetailsSummary(): SummaryListItem[] {
+    const dateAndTimeSummary = [
+      {
+        key: 'Date',
+        lines: [DateUtils.getDateStringFromDateTimeString(this.appointment.appointmentTime)],
+      },
+      {
+        key: 'Time',
+        lines: [DateUtils.getTimeStringFromDateTimeString(this.appointment.appointmentTime)],
+      },
+    ]
+
+    if (this.assignedCaseworker) {
+      return [
+        {
+          key: 'Caseworker',
+          lines: [this.assignedCaseworker.username],
+        },
+        ...dateAndTimeSummary,
+      ]
+    }
+
+    return dateAndTimeSummary
+  }
+}
